+++ napari/plugins/_plugin_manager.py
@@ -130,8 +130,11 @@ class NapariPluginManager(PluginManager):
         for spec_name, hook_caller in self.hooks.items():
             order = []
             for p in new_order.get(spec_name, []):
-                order.append(p['plugin'])
-                hook_caller._set_plugin_enabled(p['plugin'], p['enabled'])
+                try:
+                    hook_caller._set_plugin_enabled(p['plugin'], p['enabled'])
+                    order.append(p['plugin'])
+                except KeyError:
+                    pass
             if order:
                 hook_caller.bring_to_front(order)
