+++ napari/utils/settings/__init__.py
@@ -3,3 +3,5 @@
 
 from ._defaults import CORE_SETTINGS
 from ._manager import SETTINGS
+
+__all__ = ['SETTINGS', 'CORE_SETTINGS']
